(tests,refactor) Remove likely dead test code: sock_accept won't raise in tests
There could be some pathalogical case where we should deal with the case where
sock_accept can raise exceptions, but opting to not do have it for the test
server for simplicity and to not have untested code in coverage.

Note that sock_accept raising an exception initially will still propagate. Only
ones after BlockingIOError will now not propagate.